Post and Category pages

+++ codes/resources/js/pages/admin/Brands.vue
@@ -260,6 +260,9 @@ export default {
                 })
         },
         deleteBrand(id) {
+           if (!confirm("Are you sure you want to delete ?")) {
+              return false;
+           }
             axios.delete('/admin/brand/' + id)
                 .then(res => {
                     this.show_toast(res.data.status, res.data.msg);
