+++ codes/resources/js/pages/admin/Colors.vue
@@ -248,6 +248,9 @@ export default {
         })
     },
     deleteColor(id) {
+       if (!confirm("Are you sure you want to delete ?")) {
+          return false;
+       }
       axios.delete('/admin/color/' + id)
         .then(res => {
           this.show_toast(res.data.status, res.data.msg);
