+++ codes/resources/js/pages/admin/Sizes.vue
@@ -229,6 +229,9 @@ export default {
         })
     },
     deleteSize(id) {
+       if (!confirm("Are you sure you want to delete ?")) {
+          return false;
+       }
       axios.delete('/admin/size/' + id)
         .then(res => {
           this.show_toast(res.data.status, res.data.msg);
