+++ codes/resources/js/pages/admin/Styles.vue
@@ -247,6 +247,9 @@ export default {
                 })
         },
         deleteStyle(id) {
+           if (!confirm("Are you sure you want to delete ?")) {
+              return false;
+           }
             axios.delete('/admin/style/' + id)
                 .then(res => {
                     this.show_toast(res.data.status, res.data.msg);
